Create Response_and_habituation_to_acoustic_startle_assay_zf
Habituation to acoustic startle is a measure of sensorimotor gating and is sensitive to modulation by dopaminergic agonists and antagonists (Quednow et al., 2006; Burgess and Granato, 2007; Halberstadt and Geyer, 2009).

The assay is performed on 5 days post fertilization zebrafish larvae using the DanioVision observation chamber. Larvae are placed in 24 multiwell plates individually, and exposed to repetetive acoustic stimuli (taps). The amplitude of the startle response represents a measure of the habituation to repetitive stimulation.